work on memeber list commit
